IGAPP-1152: finished export of non-recurring event
