+++ api-client/src/utils/eventExport.ts
@@ -1,15 +1,13 @@
 import moment, { Moment } from 'moment'
 import { EventModel } from '../index'
-import parseHTML from './parseHTML'
 
 
 const formatDate = (date: Moment): string => `${date.format('YYYYMMDDTHHmm')}00Z`
 
-// TODO new CalendarEvent with the $recurringEvent set should have the same UID as its recurring event.
 const generateUID = (): string => {
   let generatedID = ''
   const length = 20
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
+  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
   for (let i = 0; i < length; i += 1) {
     generatedID += characters.charAt(Math.floor(Math.random() * characters.length))
   }
@@ -17,12 +15,10 @@ const generateUID = (): string => {
   return `${generatedID}@integreat.de`
 }
 
-// TODO check formatting
 export const mapToICalFormat = (event: EventModel): string => {
   const body: string[] = []
   body.push(`DTSTAMP:${formatDate(moment())}`)
-  // body.push(`UID:${generateUID()}`)
-  body.push('UID:ff808181-00000003@example.com')
+  body.push(`UID:${generateUID()}`)
   body.push(`SUMMARY:${event.title}`)
   body.push(`DTSTART:${formatDate(event.date.startDate)}`)
   body.push(`DEND:${formatDate(event.date.endDate)}`)
@@ -30,14 +26,12 @@ export const mapToICalFormat = (event: EventModel): string => {
     body.push(`LOCATION:${event.location.fullAddress}`)
   }
 
-  // TODO escape html?
-  // https://www.npmjs.com/package/html-to-text
-  if (event.content) {
-    body.push(`DESCRIPTION:${parseHTML(event.content)}`)
+  // we can't include the actual content because in some cases it can lead to an import error (not trivial to fix).
+  // That's why we link to the website instead
+  if (event.path) {
+    body.push(`DESCRIPTION:https://integreat.app${event.path}`)
   }
 
-
-  // TODO check if \n is ok
   return [
     'BEGIN:VCALENDAR',
     'VERSION:2.0',
