+++ web/src/components/Page.tsx
@@ -7,6 +7,7 @@ import DateFormatter from 'api-client/src/i18n/DateFormatter'
 import Caption from './Caption'
 import LastUpdateInfo from './LastUpdateInfo'
 import RemoteContent from './RemoteContent'
+import TextButton from './TextButton'
 
 export const THUMBNAIL_WIDTH = 300
 
@@ -30,23 +31,27 @@ type PageProps = {
   formatter: DateFormatter
   onInternalLinkClick: (url: string) => void
   children?: ReactNode
+  buttonText?: string
+  onButtonClick?: () => void
 }
 
 /**
  * Display a single page with all necessary information
  */
 const Page = ({
-  title,
-  defaultThumbnailSrc,
-  thumbnailSrcSet,
-  content,
-  lastUpdate,
-  showLastUpdateText = true,
-  lastUpdateFormat,
-  formatter,
-  children,
-  onInternalLinkClick,
-}: PageProps): ReactElement => (
+                title,
+                defaultThumbnailSrc,
+                thumbnailSrcSet,
+                content,
+                lastUpdate,
+                showLastUpdateText = true,
+                lastUpdateFormat,
+                formatter,
+                children,
+                onInternalLinkClick,
+                buttonText,
+                onButtonClick
+              }: PageProps): ReactElement => (
   <>
     {!!defaultThumbnailSrc && <Thumbnail alt='' src={defaultThumbnailSrc} srcSet={thumbnailSrcSet} />}
     <Caption title={title} />
@@ -60,6 +65,7 @@ const Page = ({
         withText={showLastUpdateText}
       />
     )}
+    {onButtonClick && !!buttonText && <TextButton onClick={onButtonClick} text={buttonText} />}
   </>
 )
